+++ srcs/main.c
@@ -1,25 +1,26 @@
 #include "fractol.h"
 
+/*
 void end(void)__attribute__((destructor));
 
 void end(void)
 {
     system("leaks fractol");
 }
-
+*/
 void	ft_update_fractal(t_data *data)
 {
 	int		x;
 	int		y;
-	double	x_cp;
-	double	y_cp;
+	long double	x_cp;
+	long double	y_cp;
 
 	mlx_mouse_get_pos(data->mlx, data->win, &x, &y);
 	//	if (!x && !y)
 	if (!x || !y)
 		return ;
-	x_cp = ((data->max_r - data->min_r) * (double)x) / WIDTH + data->min_r;
-	y_cp = ((data->max_i - data->min_i) * (double)y) / HEIGHT + data->min_i;
+	x_cp = ((data->max_r - data->min_r) * (long double)x) / (long double)WIDTH + data->min_r;
+	y_cp = ((data->max_i - data->min_i) * (long double)y) / (long double)HEIGHT + data->min_i;
 	data->c_r = x_cp;
 	data->c_i = y_cp;
 }
